[RHTAP-1321] Removed unused methods
